+++ g2p_programs/CHANGELOG.md
@@ -2,6 +2,7 @@ Format in CHANGELOG.md
 
 ## Version[0.06] - 2022-05-19
 
+- [FIX] Job Queue to import eligible beneficiaries 
 - [ADD] Notification, Program, and Cycle Managers model and UI
 - [ADD] Entitlement Manager model and UI